Correct description of "cnf" use for PoP
Let's assume we're talking about what the AS puts *in* the token
(not the token response).  That means that "cnf" always has to
be present.  The flow where "cnf" is only required in the symmetric
case and is optional for the asymmetric case is the use of the
"cnf" parameter in the token response, from AS to client.